+++ db.json
@@ -35,6 +35,12 @@
       "name": "도스타코스 선릉점",
       "description": "멕시칸 캐주얼 그릴",
       "category": "기타"
+    },
+    {
+      "id": "4a0715a0-de5a-41a3-a986-d705de0dacfa",
+      "name": "asdf",
+      "description": "asdf",
+      "category": "중식"
     }
   ]
 }
