+++ src/components/Aside/AddRestaurantModal.jsx
@@ -2,8 +2,8 @@ import Categories from "../../data/Category.js";
 import Modal from "./Modal.jsx"
 import styled from "styled-components";
 import { useSetRecoilState} from "recoil";
-import {modalState} from "../../recoil/ModalState.jsx";
-import {useRestaurants} from "../../hooks/useRestaurants.js";
+import { modalState } from "../../recoil/ModalState.jsx";
+import { addRestaurant } from "../../hooks/addRestaurant.js";
 import {v4 as uuidv4} from "uuid";
 
 const FormItem = styled.div`
@@ -83,7 +83,7 @@ const Button = styled.button`
 `;
 
 function AddRestaurantModal() {
-    const { addRestaurant } = useRestaurants();
+    const { addNewRestaurant } = addRestaurant();
     const setIsModalOpen = useSetRecoilState(modalState);
 
     const categories = Categories().filter(
@@ -102,12 +102,13 @@ function AddRestaurantModal() {
         const formData = new FormData(event.target);
         const formJson = Object.fromEntries(formData.entries());
 
-        addRestaurant({
+        addNewRestaurant({
             id: uuidv4(),
             name: formJson.name,
             description: formJson.description,
             category: formJson.category
         });
+
         handleClose();
     }
