+++ src/hooks/getRestaurants.js
@@ -1,0 +1,23 @@
+import { useEffect } from 'react';
+import { useRecoilState } from "recoil";
+import { restaurantListState } from "../recoil/RestaurantListState.jsx";
+
+export const getRestaurants = () => {
+    const [ restaurantList, setRestaurantList ] = useRecoilState(restaurantListState);
+      
+    const fetchRestaurants = async () => {
+        try {
+            const response = await fetch("http://localhost:3000/restaurants");
+            const data = await response.json();
+            setRestaurantList(data);
+        } catch (error) {
+            alert("레스토랑 데이터를 가져오는데 문제가 발생했습니다.\n" + error);
+        }
+    };
+
+    useEffect(() => {
+        fetchRestaurants();
+      }, []);
+
+    return { restaurantList };
+};
